Created project and made a start
The majority of the work has now been completed for the login server, I also setup the database tables required to query the information and started working on the clients role in communicating between the two.

Next steps are to finish working on the client, then complete the SendData class on the login server and begin testing.