Login to Synchronization server connectivty is working as well as authentication
Started implementing the synchronization server and the cache, the server successfully connects to the login server without errors and authenticates itself using the database.

I also began implementing the cache data structure, using dictionaries to store database table data. This will only cache a selected number of tables with highly volatile and constantly updating data such as position coordinates, with the other lesser commands being stored as SQL queries and batch processed when synchronization happens.

Next steps on this are to continue with implementation once the login server is finished.